times/requirements.txt
anyio==3.6.2
appdirs==1.4.4
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.0
boto3==1.26.102
botocore==1.29.102
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
fastapi==0.88.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
lxml==4.9.2
numpy==1.24.0
outcome==1.2.0
packaging==23.0
pandas==1.5.2
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.2
s3transfer==0.6.0
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.4
starlette==0.25.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.4.0
urllib3==1.26.15
uvicorn==0.20.0
webdriver-manager==3.8.5
webdrivermanager==0.10.0
wsproto==1.2.0
